BE-Rafif-003 : Update user profile
