+++ controllers/index.js
@@ -1,9 +1,7 @@
 const userControllers = require("./userControllers");
 const productController = require("./productController");
-const profileController = require("./profileController");
 
 module.exports = {
   userControllers,
   productController,
-  profileController
 };
