+++ routers/index.js
@@ -1,9 +1,7 @@
 const userRouters = require("./userRouter");
 const productRouter = require("./productRouter");
-const profileRouter = require("./profileRouter");
 
 module.exports = {
   userRouters,
   productRouter,
-  profileRouter
 };
